feat: show floored value, add inaccuracy note
in certain cases, the formula for step calculation yields values that are slightly off from the actual value shown in the game.
e.g. it was reported that a PM on Fairytale PST (cc 1.0) with Tsumugi at 16 step stat yields 2.2 progress instead of 2.1.

the calculation will be corrected when we have more information, but so far we have no clue.